requirements_uploader.txt
-r requirements.txt
discord.py
requests
